+++ src/Hydrogen/ClusteredStorage/IClusteredStorage.cs
@@ -15,38 +15,38 @@ namespace Hydrogen;
 public interface IClusteredStorage : ILoadable {
 
 	event EventHandlerEx<ClusteredStreamRecord> RecordCreated;
-	event EventHandlerEx<int, ClusteredStreamRecord> RecordAdded;
-	event EventHandlerEx<int, ClusteredStreamRecord> RecordInserted;
-	event EventHandlerEx<int, ClusteredStreamRecord> RecordUpdated;
-	event EventHandlerEx<int> RecordRemoved;
+	event EventHandlerEx<long, ClusteredStreamRecord> RecordAdded;
+	event EventHandlerEx<long, ClusteredStreamRecord> RecordInserted;
+	event EventHandlerEx<long, ClusteredStreamRecord> RecordUpdated;
+	event EventHandlerEx<long> RecordRemoved;
 
 	ClusteredStorageHeader Header { get; }
 
 	ClusteredStoragePolicy Policy { get; }
 
-	int Count { get; }
+	long Count { get; }
 
 	IReadOnlyList<ClusteredStreamRecord> Records { get; }
 
 	Endianness Endianness { get; }
 
-	ClusteredStreamRecord GetRecord(int index);
+	ClusteredStreamRecord GetRecord(long index);
 
 	ClusteredStreamScope Add();
 
-	ClusteredStreamScope Open(int index);
+	ClusteredStreamScope Open(long index);
 
-	void Remove(int index);
+	void Remove(long index);
 
-	ClusteredStreamScope Insert(int index);
+	ClusteredStreamScope Insert(long index);
 
-	void Swap(int first, int second);
+	void Swap(long first, long second);
 
 	void Clear();
 
-	ClusteredStreamScope EnterSaveItemScope<TItem>(int index, TItem item, IItemSerializer<TItem> serializer, ListOperationType operationType);
+	ClusteredStreamScope EnterSaveItemScope<TItem>(long index, TItem item, IItemSerializer<TItem> serializer, ListOperationType operationType);
 
-	ClusteredStreamScope EnterLoadItemScope<TItem>(int index, IItemSerializer<TItem> serializer, out TItem item);
+	ClusteredStreamScope EnterLoadItemScope<TItem>(long index, IItemSerializer<TItem> serializer, out TItem item);
 
 	internal IDisposable EnterLockScope();
 }
@@ -54,12 +54,12 @@ public interface IClusteredStorage : ILoadable {
 
 public static class IClusteredStorageExtensions {
 
-	public static bool IsNull(this IClusteredStorage storage, int index) {
+	public static bool IsNull(this IClusteredStorage storage, long index) {
 		return storage.GetRecord(index).Traits.HasFlag(ClusteredStreamTraits.IsNull);
 	}
 
 
-	public static byte[] ReadAll(this IClusteredStorage storage, int index) {
+	public static byte[] ReadAll(this IClusteredStorage storage, long index) {
 		using var scope = storage.Open(index);
 		return scope.Stream.ReadAll();
 	}
@@ -69,19 +69,19 @@ public static class IClusteredStorageExtensions {
 		scope.Stream.Write(bytes);
 	}
 
-	public static void UpdateBytes(this IClusteredStorage storage, int index, ReadOnlySpan<byte> bytes) {
+	public static void UpdateBytes(this IClusteredStorage storage, long index, ReadOnlySpan<byte> bytes) {
 		using var scope = storage.Open(index);
 		scope.Stream.SetLength(0);
 		scope.Stream.Write(bytes);
 	}
 
-	public static void AppendBytes(this IClusteredStorage storage, int index, ReadOnlySpan<byte> bytes) {
+	public static void AppendBytes(this IClusteredStorage storage, long index, ReadOnlySpan<byte> bytes) {
 		using var scope = storage.Open(index);
 		scope.Stream.Seek(scope.Stream.Length, SeekOrigin.Current);
 		scope.Stream.Write(bytes);
 	}
 
-	public static void InsertBytes(this IClusteredStorage storage, int index, ReadOnlySpan<byte> bytes) {
+	public static void InsertBytes(this IClusteredStorage storage, long index, ReadOnlySpan<byte> bytes) {
 		using var scope = storage.Insert(index);
 		if (bytes != null) {
 			scope.Stream.Seek(scope.Stream.Length, SeekOrigin.Current);
@@ -89,11 +89,11 @@ public static class IClusteredStorageExtensions {
 		}
 	}
 
-	public static void SaveItem<TItem>(this IClusteredStorage storage, int index, TItem item, IItemSerializer<TItem> serializer, ListOperationType operationType) {
+	public static void SaveItem<TItem>(this IClusteredStorage storage, long index, TItem item, IItemSerializer<TItem> serializer, ListOperationType operationType) {
 		using var scope = storage.EnterSaveItemScope(index, item, serializer, operationType);
 	}
 
-	public static TItem LoadItem<TItem>(this IClusteredStorage storage, int index, IItemSerializer<TItem> serializer) {
+	public static TItem LoadItem<TItem>(this IClusteredStorage storage, long index, IItemSerializer<TItem> serializer) {
 		using var scope = storage.EnterLoadItemScope(index, serializer, out var item);
 		return item;
 	}
