+++ src/Hydrogen/Collections/Collections/IExtendedCollection.cs
@@ -8,18 +8,13 @@
 
 using System.Collections.Generic;
 
-namespace Hydrogen;
+namespace Hydrogen; 
 
 public interface IExtendedCollection<T> : ICollection<T>, IReadOnlyExtendedCollection<T>, IWriteOnlyExtendedCollection<T> {
-	new int Count { get; }
-
+	new long Count { get; }
 	new void Add(T item);
-
 	new void Clear();
-
 	new bool Contains(T item);
-
 	new void CopyTo(T[] array, int arrayIndex);
-
 	new bool Remove(T item);
 }
