+++ src/Hydrogen/Collections/Lists/BoundedList.cs
@@ -18,34 +18,34 @@ public class BoundedList<T> : ExtendedListDecorator<T>, IBoundedList<T> {
 		FirstIndex = startIndex;
 	}
 
-	public int FirstIndex { get; }
+	public long FirstIndex { get; }
 
-	public override IEnumerable<int> IndexOfRange(IEnumerable<T> items) {
+	public override IEnumerable<long> IndexOfRange(IEnumerable<T> items) {
 		var startIX = FirstIndex;
 		return base.IndexOfRange(items).Select(x => x + startIX);
 	}
 
-	public override IEnumerable<T> ReadRange(int index, int count) {
+	public override IEnumerable<T> ReadRange(long index, long count) {
 		CheckRange(index, count);
 		return base.ReadRange(index - FirstIndex, count);
 	}
 
-	public override void UpdateRange(int index, IEnumerable<T> items) {
+	public override void UpdateRange(long index, IEnumerable<T> items) {
 		CheckRange(index, items.Count());
 		base.UpdateRange(index - FirstIndex, items);
 	}
 
-	public override void InsertRange(int index, IEnumerable<T> items) {
+	public override void InsertRange(long index, IEnumerable<T> items) {
 		CheckRange(index, 0);
 		base.InsertRange(index - FirstIndex, items);
 	}
 
-	public override void RemoveRange(int index, int count) {
+	public override void RemoveRange(long index, long count) {
 		CheckRange(index, count);
 		base.RemoveRange(index - FirstIndex, count);
 	}
 
-	protected void CheckRange(int index, int count) {
+	protected void CheckRange(long index, long count) {
 		var startIX = FirstIndex;
 		var lastIX = startIX + (base.Count - 1).ClipTo(startIX, int.MaxValue);
 		Guard.ArgumentInRange(index, startIX, lastIX, nameof(index));
