+++ src/Hydrogen/Collections/Lists/IReadOnlyBoundedList.cs
@@ -9,5 +9,5 @@
 namespace Hydrogen;
 
 public interface IReadOnlyBoundedList<T> : IReadOnlyExtendedList<T> {
-	int FirstIndex { get; }
+	long FirstIndex { get; }
 }
